Move BoundaryData and FLuxBC to base
