+++ cpp/dolfinx_eqlb/KernelData.hpp
@@ -302,239 +302,4 @@ protected:
   std::vector<double> _data_transform_shpfkt;
 };
 
-template <typename T>
-class KernelDataBC : public base::KernelData<T>
-{
-public:
-  /// Kernel data for calculation of patch boundary-conditions
-  ///
-  /// Holds the required surface-quadrature rule, calculates required mappings,
-  /// interpolates from a flux function and calculates DOFs of a flux-function.
-  ///
-  /// @param[in] mesh                 The mesh
-  /// @param[in] quadrature_rule_fct  The quadrature rule on the cell facets
-  /// @param[in] basix_element_fluxpw The basix-element for the H(div) flux
-  /// @param[in] basix_element_rhs    The basix-element for RHS and proj. flux
-  KernelDataBC(std::shared_ptr<const mesh::Mesh> mesh,
-               std::shared_ptr<const base::QuadratureRule> quadrature_rule_fct,
-               std::shared_ptr<const fem::FiniteElement> element_flux_hdiv,
-               const int nfluxdofs_per_fct, const int nfluxdofs_cell,
-               const bool flux_is_custom);
-
-  /* Tabulate/map shape-function sof the flux-space */
-
-  /// Tabulates a basix_element at the surface-quadrature points
-  /// @param[in] basix_element The Basix element
-  /// @param[in,out] storage   Stoarge vector for the tabulates shape-functions
-  /// @param[out] shape_vector The shape for creating an mdspan of the tabulated
-  /// functions
-  std::array<std::size_t, 5>
-  shapefunctions_flux_qpoints(const basix::FiniteElement& basix_element,
-                              std::vector<double>& storage)
-  {
-    return this->tabulate_basis(basix_element,
-                                this->_quadrature_rule[0]->points(), storage,
-                                false, false);
-  }
-
-  /// Map flux-functions from refrence to current cell
-  ///
-  /// Aplies the contra-variant Piola mapping to the shape-functions of the
-  /// flux. The here performed mapping is restriced to the shape-functions of
-  /// one cell facet. All other functions will be neglected.
-  ///
-  /// @param[in] lfct_id      The cell-local Id of the facet
-  /// @param[in, out] phi_cur The shape-function on the current cell
-  /// @param[in] phi_ref      The shape-functions on the reference cell
-  /// @param[in] J            The Jacobina
-  /// @param[in] detJ         The determinant of the Jacobian
-  void map_shapefunctions_flux(std::int8_t lfct_id,
-                               base::mdspan_t<double, 3> phi_cur,
-                               base::mdspan_t<const double, 5> phi_ref,
-                               base::mdspan_t<const double, 2> J, double detJ);
-
-  /* Calculate flux DOFs based on different inputs */
-
-  /// Calculates a flux (vector) from a given normal-trace on a cell-facet
-  ///
-  /// Takes over a list of normal-traces at (multiple) points and return the
-  /// flux-vectors calculates from facet_normal x trace, where the facte_normal
-  /// has a magnitude of 1.
-  ///
-  /// @param[in] flux_ntrace_cur The vector of flux normal-traces within the
-  ///                            current cell
-  /// @param[in] lfct_id         The cell-local Id of the facet
-  /// @param[in] K               The inverse of the Jacobian
-  /// @param[out] flux_vector    A List of flux-vectors recovered from the
-  ///                            normal-trace
-  base::mdspan_t<const T, 2>
-  normaltrace_to_flux(std::span<const T> flux_ntrace_cur, std::int8_t lfct_id,
-                      base::mdspan_t<const double, 2> K)
-  {
-    // Perform calculation
-    normaltrace_to_vector(flux_ntrace_cur, lfct_id, K);
-
-    return base::mdspan_t<const T, 2>(_flux_scratch_data.data(),
-                                      flux_ntrace_cur.size(),
-                                      (std::size_t)this->_gdim);
-  }
-
-  /// Calculates flux DOFs based on the normal-trace
-  ///
-  /// Claculates flux DOFs on a cell facet based on given normal-trace,
-  /// evaluated at the required interpolation points of the flux-space.
-  ///
-  /// @param[in] flux_ntrace_cur The flux normal-trace on the interpolation
-  ///                            points
-  /// @param[in, out] flux_dofs  The calculated flux DOFs on the facet
-  /// @param[in] lfct_id         The (cell-local) Id of the facet
-  /// @param[in] J               The Jacobi matrix of the mapping function
-  /// @param[in] detJ            The determinant of the Jacobi matrix
-  /// @param[in] K               The inverse of the Jacobi matrix
-  void interpolate_flux(std::span<const T> flux_ntrace_cur,
-                        std::span<T> flux_dofs, std::int8_t lfct_id,
-                        base::mdspan_t<const double, 2> J, double detJ,
-                        base::mdspan_t<const double, 2> K);
-
-  /// Calculates boundary DOFs for a patch problem
-  ///
-  /// Calculates the boundary DOFs for a patrch problem from the boundary DOFs
-  /// where the hat-function is neglected by inteprolating boundary_function x
-  /// hat_function.
-  ///
-  /// @param[in] flux_dofs_bc         The boundary DOFs without consideration of
-  ///                                 the hat-function
-  /// @param[in] cell_id              Vector with Ids of the boundary cells
-  /// @param[in] lfct_id              Vector with the (cell-local) Ids of the
-  ///                                 boundary facets
-  /// @param[in] hat_id               Vector of (cell-local) Ids of the
-  ///                                 patch-central node
-  /// @param[in] cell_info            The informatios required for DOF
-  ///                                 transformations on a cell
-  /// @param[in] J                    The Jacobi matrix of the mapping function
-  /// @param[in] detJ                 The determinant of the Jacobi matrix
-  /// @param[in] K                    The inverse of the Jacobi matrix
-  /// @param[out] flux_dofs_patch     The boundary DOFs for a patch-problem
-  std::vector<T> interpolate_flux(std::span<const T> flux_dofs_bc,
-                                  std::int32_t cell_id, std::int8_t lfct_id,
-                                  std::int8_t hat_id,
-                                  std::span<const std::uint32_t> cell_info,
-                                  base::mdspan_t<const double, 2> J,
-                                  double detJ,
-                                  base::mdspan_t<const double, 2> K)
-  {
-    // Initialise storage
-    std::vector<T> flux_dofs_patch(flux_dofs_bc.size());
-
-    // Interpolaate flux
-    interpolate_flux(flux_dofs_bc, flux_dofs_patch, cell_id, lfct_id, hat_id,
-                     cell_info, J, detJ, K);
-
-    return std::move(flux_dofs_patch);
-  }
-
-  /// Calculates boundary DOFs for a patch problem
-  ///
-  /// Calculates the boundary DOFs for a patrch problem from the boundary DOFs
-  /// where the hat-function is neglected by inteprolating boundary_function x
-  /// hat_function.
-  ///
-  /// This routine should be called where performace is relevant!
-  ///
-  /// @param[in] flux_dofs_bc         The boundary DOFs without consideration of
-  ///                                 the hat-function
-  /// @param[in, out] flux_dofs_patch The boundary DOFs for a patch-problem
-  /// @param[in] cell_id              Vector with Ids of the boundary cells
-  /// @param[in] lfct_id              Vector with the (cell-local) Ids of the
-  ///                                 boundary facets
-  /// @param[in] hat_id               Vector of (cell-local) Ids of the
-  ///                                 patch-central node
-  /// @param[in] cell_info            The informatios required for DOF
-  ///                                 transformations on a cell
-  /// @param[in] J                    The Jacobi matrix of the mapping function
-  /// @param[in] detJ                 The determinant of the Jacobi matrix
-  /// @param[in] K                    The inverse of the Jacobi matrix
-  void interpolate_flux(std::span<const T> flux_dofs_bc,
-                        std::span<T> flux_dofs_patch, std::int32_t cell_id,
-                        std::int8_t lfct_id, std::int8_t hat_id,
-                        std::span<const std::uint32_t> cell_info,
-                        base::mdspan_t<const double, 2> J, double detJ,
-                        base::mdspan_t<const double, 2> K);
-
-  /// Calculates flux-DOFs from flux-vector at the interpolation points
-  ///
-  /// Calculates flux DOFs on one cell facte from the flux-values (on the
-  /// current cell) on the interpolation points of the flux-space.
-  ///
-  /// @param[in] flux_cur             The flux (current cell) on the
-  ///                                 interpolation points
-  /// @param[in, out] flux_dofs_patch The calculated flux DOFs
-  /// @param[in] lfct_id              The (cell-local) Id of the facet on which
-  ///                                 the DOFs are calculated
-  /// @param[in] J                    The Jacobi matrix of the mapping function
-  /// @param[in] detJ                 The determinant of the Jacobi matrix
-  /// @param[in] K                    The inverse of the Jacobi matrix
-  void interpolate_flux(base::mdspan_t<const T, 2> flux_cur,
-                        std::span<T> flux_dofs_patch, std::int8_t lfct_id,
-                        base::mdspan_t<const double, 2> J, double detJ,
-                        base::mdspan_t<const double, 2> K);
-
-  /* Getter methods: Interpolation */
-
-  /// Extract number of interpolation points
-  /// @return Number of interpolation points
-  int num_interpolation_points() const { return _nipoints; }
-
-  /// Extract number of interpolation points per facet
-  /// @return Number of interpolation points
-  int num_interpolation_points_per_facet() const { return _nipoints_per_fct; }
-
-protected:
-  void normaltrace_to_vector(std::span<const T> normaltrace_cur,
-                             std::int8_t lfct_id,
-                             base::mdspan_t<const double, 2> K);
-
-  /* Variable definitions */
-
-  // Interpolation data
-  std::size_t _nipoints_per_fct, _nipoints;
-  std::vector<double> _ipoints, _data_M;
-  base::mdspan_t<const double, 4> _M; // Indices: facet, dof, gdim, points
-
-  // Tabulated shape-functions H(div) flux (integration points)
-  std::vector<double> _basis_flux_values;
-  base::mdspan_t<const double, 5> _basis_flux;
-  const int _ndofs_per_fct, _ndofs_fct, _ndofs_cell;
-
-  // Tabulated shape-functions (hat-function)
-  basix::FiniteElement _basix_element_hat;
-  std::vector<double> _basis_hat_values;
-  base::mdspan_t<const double, 5> _basis_hat;
-
-  // Pull-back H(div) data
-  std::size_t _size_flux_scratch;
-  std::vector<T> _flux_scratch_data, _mflux_scratch_data;
-  base::mdspan_t<T, 2> _flux_scratch, _mflux_scratch;
-
-  std::array<double, 3> _normal_scratch;
-
-  std::function<void(base::mdspan_t<T, 2>&, const base::mdspan_t<const T, 2>&,
-                     const base::mdspan_t<const double, 2>&, double,
-                     const base::mdspan_t<const double, 2>&)>
-      _pull_back_flux;
-
-  // Push-forward H(div) shape-functions
-  std::vector<double> _mbasis_flux_values, _mbasis_scratch_values;
-  base::mdspan_t<double, 2> _mbasis_flux, _mbasis_scratch;
-
-  std::function<void(base::mdspan_t<double, 2>&,
-                     const base::mdspan_t<const double, 2>&,
-                     const base::mdspan_t<const double, 2>&, double,
-                     const base::mdspan_t<const double, 2>&)>
-      _push_forward_flux;
-
-  std::function<void(const std::span<double>&,
-                     const std::span<const std::uint32_t>&, std::int32_t, int)>
-      _apply_dof_transformation;
-};
 } // namespace dolfinx_eqlb
