+++ cpp/dolfinx_eqlb/base/BoundaryData.cpp
@@ -7,8 +7,274 @@
 #include "BoundaryData.hpp"
 
 using namespace dolfinx;
-using namespace dolfinx_eqlb;
+using namespace dolfinx_eqlb::base;
+// ------------------------------------------------------------------------------
+/* KernelDataBC */
+// ------------------------------------------------------------------------------
+template <typename T>
+KernelDataBC<T>::KernelDataBC(
+    std::shared_ptr<const mesh::Mesh> mesh,
+    std::shared_ptr<const QuadratureRule> quadrature_rule_fct,
+    std::shared_ptr<const fem::FiniteElement> element_flux_hdiv,
+    const int nfluxdofs_per_fct, const int nfluxdofs_cell,
+    const bool flux_is_custom)
+    : KernelData<T>(mesh, {quadrature_rule_fct}),
+      _ndofs_per_fct(nfluxdofs_per_fct),
+      _ndofs_fct(this->_nfcts_per_cell * _ndofs_per_fct),
+      _ndofs_cell(nfluxdofs_cell),
+      _basix_element_hat(basix::element::create_lagrange(
+          mesh::cell_type_to_basix_type(mesh->topology().cell_type()), 1,
+          basix::element::lagrange_variant::equispaced, false))
+{
+  // Extract the baisx flux element
+  const basix::FiniteElement& basix_element_flux_hdiv
+      = element_flux_hdiv->basix_element();
 
+  /* Interpolation points on facets */
+  std::array<std::size_t, 4> shape_intpl = this->interpolation_data_facet_rt(
+      basix_element_flux_hdiv, flux_is_custom, this->_gdim,
+      this->_nfcts_per_cell, _ipoints, _data_M);
+
+  _M = mdspan_t<const double, 4>(_data_M.data(), shape_intpl);
+
+  std::tie(_nipoints_per_fct, _nipoints)
+      = this->size_interpolation_data_facet_rt(shape_intpl);
+
+  /* Tabulate required shape-functions */
+  std::array<std::size_t, 5> shape;
+
+  // Tabulate H(div) flux at interpolation points
+  shape = this->tabulate_basis(basix_element_flux_hdiv, _ipoints,
+                               _basis_flux_values, false, false);
+  _basis_flux = mdspan_t<const double, 5>(_basis_flux_values.data(), shape);
+
+  _mbasis_flux_values.resize(_ndofs_cell * this->_gdim);
+  _mbasis_scratch_values.resize(_ndofs_cell * this->_gdim);
+  _mbasis_flux = mdspan_t<double, 2>(_mbasis_flux_values.data(), _ndofs_cell,
+                                     this->_gdim);
+  _mbasis_scratch = mdspan_t<double, 2>(_mbasis_scratch_values.data(),
+                                        _ndofs_cell, this->_gdim);
+
+  // Tabulate hat-function at interpolation points
+  shape = this->tabulate_basis(_basix_element_hat, _ipoints, _basis_hat_values,
+                               false, false);
+  _basis_hat = mdspan_t<const double, 5>(_basis_hat_values.data(), shape);
+
+  /* H(div)-flux: Pull back into reference */
+  // Initialise scratch
+  _size_flux_scratch
+      = std::max(_nipoints_per_fct, this->_quadrature_rule[0]->num_points(0));
+  _flux_scratch_data.resize(_size_flux_scratch * this->_gdim);
+  _mflux_scratch_data.resize(_size_flux_scratch * this->_gdim);
+
+  _flux_scratch = mdspan_t<T, 2>(_flux_scratch_data.data(), _size_flux_scratch,
+                                 this->_gdim);
+  _mflux_scratch = mdspan_t<T, 2>(_mflux_scratch_data.data(),
+                                  _size_flux_scratch, this->_gdim);
+
+  /* Extract mapping functions */
+  using J_t = mdspan_t<const double, 2>;
+  using K_t = mdspan_t<const double, 2>;
+
+  // Push-forward for shape-functions
+  using u_t = mdspan_t<double, 2>;
+  using U_t = mdspan_t<const double, 2>;
+
+  _push_forward_flux = basix_element_flux_hdiv.map_fn<u_t, U_t, K_t, J_t>();
+
+  // Pull-back for values
+  using V_t = mdspan_t<T, 2>;
+  using v_t = mdspan_t<const T, 2>;
+
+  _pull_back_flux = basix_element_flux_hdiv.map_fn<V_t, v_t, K_t, J_t>();
+
+  // DOF-transformation function (shape functions)
+  _apply_dof_transformation
+      = element_flux_hdiv->get_dof_transformation_function<double>(false, false,
+                                                                   false);
+}
+
+template <typename T>
+void KernelDataBC<T>::map_shapefunctions_flux(std::int8_t lfct_id,
+                                              mdspan_t<double, 3> phi_cur,
+                                              mdspan_t<const double, 5> phi_ref,
+                                              mdspan_t<const double, 2> J,
+                                              double detJ)
+{
+  const int offs_pnt = lfct_id * this->_quadrature_rule[0]->num_points(0);
+  const int offs_dof = lfct_id * _ndofs_per_fct;
+
+  // Loop over all evaluation points
+  for (std::size_t i = 0; i < phi_cur.extent(0); ++i)
+  {
+    // Index of the reference shape-functions
+    int i_ref = offs_pnt + i;
+
+    // Loop over all basis functions
+    for (std::size_t j = 0; j < phi_cur.extent(1); ++j)
+    {
+      // Index of the reference shape-functions
+      int j_ref = offs_dof + j;
+
+      // Inverse of the determinenant
+      double inv_detJ = 1.0 / detJ;
+
+      // Evaluate (1/detj) * J * phi^j(x_i)
+      double acc = 0;
+
+      if (phi_cur.extent(2) == 2)
+      {
+        acc = J(0, 0) * phi_ref(0, 0, i_ref, j_ref, 0)
+              + J(0, 1) * phi_ref(0, 0, i_ref, j_ref, 1);
+        phi_cur(i, j, 0) = inv_detJ * acc;
+
+        acc = J(1, 0) * phi_ref(0, 0, i_ref, j_ref, 0)
+              + J(1, 1) * phi_ref(0, 0, i_ref, j_ref, 1);
+        phi_cur(i, j, 1) = inv_detJ * acc;
+      }
+      else
+      {
+        acc = J(0, 0) * phi_ref(0, 0, i_ref, j_ref, 0)
+              + J(0, 1) * phi_ref(0, 0, i_ref, j_ref, 1)
+              + J(0, 2) * phi_ref(0, 0, i_ref, j_ref, 2);
+        phi_cur(i, j, 0) = inv_detJ * acc;
+
+        acc = J(1, 0) * phi_ref(0, 0, i_ref, j_ref, 0)
+              + J(1, 1) * phi_ref(0, 0, i_ref, j_ref, 1)
+              + J(1, 2) * phi_ref(0, 0, i_ref, j_ref, 2);
+        phi_cur(i, j, 1) = inv_detJ * acc;
+
+        acc = J(2, 0) * phi_ref(0, 0, i_ref, j_ref, 0)
+              + J(2, 1) * phi_ref(0, 0, i_ref, j_ref, 1)
+              + J(2, 2) * phi_ref(0, 0, i_ref, j_ref, 2);
+        phi_cur(i, j, 1) = inv_detJ * acc;
+      }
+    }
+  }
+}
+
+template <typename T>
+void KernelDataBC<T>::interpolate_flux(std::span<const T> flux_ntrace_cur,
+                                       std::span<T> flux_dofs,
+                                       std::int8_t lfct_id,
+                                       mdspan_t<const double, 2> J, double detJ,
+                                       mdspan_t<const double, 2> K)
+{
+  // Calculate flux within current cell
+  normaltrace_to_vector(flux_ntrace_cur, lfct_id, K);
+
+  // Calculate DOFs based on values at interpolation points
+  interpolate_flux(_flux_scratch, flux_dofs, lfct_id, J, detJ, K);
+}
+
+template <typename T>
+void KernelDataBC<T>::interpolate_flux(std::span<const T> flux_dofs_bc,
+                                       std::span<T> flux_dofs_patch,
+                                       std::int32_t cell_id,
+                                       std::int8_t lfct_id, std::int8_t hat_id,
+                                       std::span<const std::uint32_t> cell_info,
+                                       mdspan_t<const double, 2> J, double detJ,
+                                       mdspan_t<const double, 2> K)
+{
+  /* Map shape functions*/
+  // Copy shape-functions from reference element
+  const int offs_ipnt = lfct_id * _nipoints_per_fct;
+  const int offs_dof = lfct_id * _ndofs_per_fct;
+
+  for (std::size_t i_pnt = 0; i_pnt < _nipoints_per_fct; ++i_pnt)
+  {
+    // Copy shape-functions at current point
+    // TODO - Check if copy on only functions on current facet is enough!
+    for (std::size_t i_dof = 0; i_dof < _ndofs_fct; ++i_dof)
+    {
+      for (std::size_t i_dim = 0; i_dim < this->_gdim; ++i_dim)
+      {
+        _mbasis_scratch(i_dof, i_dim)
+            = _basis_flux(0, 0, offs_ipnt + i_pnt, i_dof, i_dim);
+      }
+    }
+
+    // Apply dof transformation
+    _apply_dof_transformation(_mbasis_scratch_values, cell_info, cell_id,
+                              this->_gdim);
+
+    // Apply push-foreward function
+    _push_forward_flux(_mbasis_flux, _mbasis_scratch, J, detJ, K);
+
+    // Evaluate flux
+    for (std::size_t i_dim = 0; i_dim < this->_gdim; ++i_dim)
+    {
+      // Evaluate flux at current point
+      T acc = 0;
+      for (std::size_t i_dof = 0; i_dof < _ndofs_per_fct; ++i_dof)
+      {
+        acc += flux_dofs_bc[i_dof] * _mbasis_flux(offs_dof + i_dof, i_dim);
+      }
+
+      // Multiply flux with hat function
+      _flux_scratch(i_pnt, i_dim)
+          = acc * _basis_hat(0, 0, offs_ipnt + i_pnt, hat_id, 0);
+    }
+  }
+
+  // Calculate DOF of scaled flux
+  interpolate_flux(_flux_scratch, flux_dofs_patch, lfct_id, J, detJ, K);
+}
+
+template <typename T>
+void KernelDataBC<T>::interpolate_flux(mdspan_t<const T, 2> flux_cur,
+                                       std::span<T> flux_dofs,
+                                       std::int8_t lfct_id,
+                                       mdspan_t<const double, 2> J, double detJ,
+                                       mdspan_t<const double, 2> K)
+{
+  // Map flux to reference cell
+  _pull_back_flux(_mflux_scratch, flux_cur, K, 1 / detJ, J);
+
+  // Apply interpolation operator
+  for (std::size_t i = 0; i < flux_dofs.size(); ++i)
+  {
+    T dof = 0;
+    for (std::size_t j = 0; j < _nipoints_per_fct; ++j)
+    {
+      for (std::size_t k = 0; k < this->_gdim; ++k)
+      {
+        {
+          dof += _M(lfct_id, i, k, j) * _mflux_scratch(j, k);
+        }
+      }
+    }
+
+    flux_dofs[i] = dof;
+  }
+}
+
+template <typename T>
+void KernelDataBC<T>::normaltrace_to_vector(std::span<const T> normaltrace_cur,
+                                            std::int8_t lfct_id,
+                                            mdspan_t<const double, 2> K)
+{
+  // Calculate physical facet normal
+  std::span<double> normal_cur(_normal_scratch.data(), this->_gdim);
+  this->physical_fct_normal(normal_cur, K, lfct_id);
+
+  // Calculate flux within current cell
+  for (std::size_t i = 0; i < normaltrace_cur.size(); ++i)
+  {
+    int offs = this->_gdim * i;
+
+    // Set flux
+    for (std::size_t j = 0; j < this->_gdim; ++j)
+    {
+      _flux_scratch(i, j) = normal_cur[j] * normaltrace_cur[i];
+    }
+  }
+}
+// ------------------------------------------------------------------------------
+
+// ------------------------------------------------------------------------------
+/* BoundaryData */
+// ------------------------------------------------------------------------------
 template <typename T>
 BoundaryData<T>::BoundaryData(
     std::vector<std::vector<std::shared_ptr<FluxBC<T>>>>& list_bcs,
@@ -31,14 +297,14 @@ BoundaryData<T>::BoundaryData(
                 + V_flux_hdiv->dofmap()->index_map->num_ghosts()),
       _fct_to_cell(
           V_flux_hdiv->mesh()->topology().connectivity(_gdim - 1, _gdim)),
-      _quadrature_rule(base::QuadratureRule(
+      _quadrature_rule(QuadratureRule(
           V_flux_hdiv->mesh()->topology().cell_type(), quadrature_degree,
           V_flux_hdiv->mesh()->geometry().dim() - 1)),
-      _kernel_data(KernelDataBC<T>(
-          V_flux_hdiv->mesh(),
-          {std::make_shared<base::QuadratureRule>(_quadrature_rule)},
-          V_flux_hdiv->element(), _ndofs_per_fct, _ndofs_per_cell,
-          rtflux_is_custom))
+      _kernel_data(
+          KernelDataBC<T>(V_flux_hdiv->mesh(),
+                          {std::make_shared<QuadratureRule>(_quadrature_rule)},
+                          V_flux_hdiv->element(), _ndofs_per_fct,
+                          _ndofs_per_cell, rtflux_is_custom))
 {
   // Resize storage
   _num_bcfcts.resize(_num_rhs);
@@ -50,7 +316,7 @@ BoundaryData<T>::BoundaryData(
 
   std::int32_t size_fctdata = _num_rhs * _num_fcts;
   _local_fct_id.resize(_num_fcts);
-  _facet_type.resize(size_fctdata, base::PatchFacetType::internal);
+  _facet_type.resize(size_fctdata, PatchFacetType::internal);
   _offset_fctdata.resize(_num_rhs + 1);
 
   if (reconstruct_stress)
@@ -70,8 +336,8 @@ BoundaryData<T>::BoundaryData(
   std::generate(_offset_fctdata.begin(), _offset_fctdata.end(),
                 [n = 0, num_fcts]() mutable { return num_fcts * (n++); });
 
-  base::mdspan_t<double, 2> J(_data_J.data(), _gdim, _gdim);
-  base::mdspan_t<double, 2> K(_data_K.data(), _gdim, _gdim);
+  mdspan_t<double, 2> J(_data_J.data(), _gdim, _gdim);
+  mdspan_t<double, 2> K(_data_K.data(), _gdim, _gdim);
 
   /* Extract required data */
   // The mesh
@@ -121,8 +387,8 @@ BoundaryData<T>::BoundaryData(
 
   // Storage for extraction of cell geometry
   std::vector<double> coordinates_data(mesh->geometry().cmap().dim() * 3);
-  base::mdspan_t<const double, 2> coordinates(coordinates_data.data(),
-                                              mesh->geometry().cmap().dim(), 3);
+  mdspan_t<const double, 2> coordinates(coordinates_data.data(),
+                                        mesh->geometry().cmap().dim(), 3);
 
   // Storage of normal-traces/normal on boundary
   std::vector<T> values_bkernel(std::max(nipoints, nqpoints));
@@ -136,8 +402,8 @@ BoundaryData<T>::BoundaryData(
   bool initialise_projection = true;
 
   std::vector<double> basis_projection_values, mbasis_projection_values;
-  base::mdspan_t<const double, 5> basis_projection;
-  base::mdspan_t<double, 3> mbasis_projection;
+  mdspan_t<const double, 5> basis_projection;
+  mdspan_t<double, 3> mbasis_projection;
 
   // Linear equation system (projection)
   Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> A_e;
@@ -160,7 +426,7 @@ BoundaryData<T>::BoundaryData(
     for (std::int32_t fct : fct_esntbound_prime[i_rhs])
     {
       // Set facet type
-      facet_type_i[fct] = base::PatchFacetType::essnt_primal;
+      facet_type_i[fct] = PatchFacetType::essnt_primal;
     }
 
     // Handle facets with essential BCs on dual problem
@@ -261,14 +527,14 @@ BoundaryData<T>::BoundaryData(
                     _V_flux_hdiv->element()->basix_element(),
                     basis_projection_values);
 
-            basis_projection = base::mdspan_t<const double, 5>(
+            basis_projection = mdspan_t<const double, 5>(
                 basis_projection_values.data(), shape);
 
             mbasis_projection_values.resize(nqpoints_per_fct * _ndofs_per_fct
                                             * _gdim);
-            mbasis_projection = base::mdspan_t<double, 3>(
-                mbasis_projection_values.data(), nqpoints_per_fct,
-                _ndofs_per_fct, _gdim);
+            mbasis_projection
+                = mdspan_t<double, 3>(mbasis_projection_values.data(),
+                                      nqpoints_per_fct, _ndofs_per_fct, _gdim);
 
             // The equation system
             A_e.resize(_ndofs_per_fct, _ndofs_per_fct);
@@ -334,7 +600,7 @@ BoundaryData<T>::BoundaryData(
         apply_inverse_dof_transform(boundary_values, cell_info, cell, 1);
 
         // Set values and markers
-        facet_type_i[fct] = base::PatchFacetType::essnt_dual;
+        facet_type_i[fct] = PatchFacetType::essnt_dual;
         _local_fct_id[fct] = fct_loc;
 
         for (std::size_t i = 0; i < _ndofs_per_fct; ++i)
@@ -383,8 +649,8 @@ void BoundaryData<T>::calculate_patch_bc(
   std::span<const double> node_coordinates = mesh->geometry().x();
 
   std::vector<double> coordinates_data(mesh->geometry().cmap().dim() * 3);
-  base::mdspan_t<const double, 2> coordinates(coordinates_data.data(),
-                                              mesh->geometry().cmap().dim(), 3);
+  mdspan_t<const double, 2> coordinates(coordinates_data.data(),
+                                        mesh->geometry().cmap().dim(), 3);
 
   /* Evaluate BCs for all RHS */
   for (std::size_t i_fct = 0; i_fct < bound_fcts.size(); i_fct++)
@@ -403,8 +669,8 @@ void BoundaryData<T>::calculate_patch_bc(
     }
 
     // Calculate J
-    base::mdspan_t<double, 2> J(_data_J.data(), _gdim, _gdim);
-    base::mdspan_t<double, 2> K(_data_K.data(), _gdim, _gdim);
+    mdspan_t<double, 2> J(_data_J.data(), _gdim, _gdim);
+    mdspan_t<double, 2> K(_data_K.data(), _gdim, _gdim);
 
     double detJ
         = _kernel_data.compute_jacobian(J, K, _detJ_scratch, coordinates);
@@ -418,14 +684,11 @@ void BoundaryData<T>::calculate_patch_bc(
 }
 
 template <typename T>
-void BoundaryData<T>::calculate_patch_bc(const int rhs_i,
-                                         const std::int32_t fct,
-                                         const std::int8_t hat_id,
-                                         base::mdspan_t<const double, 2> J,
-                                         const double detJ,
-                                         base::mdspan_t<const double, 2> K)
+void BoundaryData<T>::calculate_patch_bc(
+    const int rhs_i, const std::int32_t fct, const std::int8_t hat_id,
+    mdspan_t<const double, 2> J, const double detJ, mdspan_t<const double, 2> K)
 {
-  if (facet_type(rhs_i)[fct] == base::PatchFacetType::essnt_dual)
+  if (facet_type(rhs_i)[fct] == PatchFacetType::essnt_dual)
   {
     // The cell
     std::int32_t cell = _fct_to_cell->links(fct)[0];
@@ -515,7 +778,9 @@ void BoundaryData<T>::boundary_dofs(const std::int32_t cell,
     }
   }
 }
+// ------------------------------------------------------------------------------
 
 // ------------------------------------------------------------------------------
+template class KernelDataBC<double>;
 template class BoundaryData<double>;
 // ------------------------------------------------------------------------------
