+++ cpp/dolfinx_eqlb/base/CMakeLists.txt
@@ -1,7 +1,10 @@
 set(HEADERS_base
+    ${CMAKE_CURRENT_SOURCE_DIR}/BoundaryData.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/FluxBC.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/KernelData.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Patch.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/QuadratureRule.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/assemble_bcs_impl.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/deqlb_base.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mdspan.hpp
     PARENT_SCOPE
@@ -9,5 +12,6 @@ set(HEADERS_base
 
 target_sources(
   dolfinx_eqlb
+  PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/BoundaryData.cpp
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/KernelData.cpp
 )
